requirements.txt
streamlit
pandas
PySastrawi